requirements_production_wp.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pyjwt==2.8.0
werkzeug==3.0.1
python-dotenv==1.0.0
ollama==0.5.2
pydantic==2.5.0
mysql-connector-python==9.4.0